exporter/requirements.txt
boto3==1.7.79
gspread==3.1.0
oauth2client==4.1.3
PyYAML==5.1